Major refactoring in massive Bundler class.
 - Moved static methods as functions into a new import-utils.ts
 - Privatized many methods, renamed and sorted alphabetically.
 - Removed unused methods.
 - Added comments to document nearly all methods.
