+++ src/bundler.ts
@@ -26,6 +26,7 @@ import {ParsedHtmlDocument} from 'polymer-analyzer/lib/html/html-document';
 import {FSUrlLoader} from 'polymer-analyzer/lib/url-loader/fs-url-loader';
 import constants from './constants';
 import * as astUtils from './ast-utils';
+import * as importUtils from './import-utils';
 import * as matchers from './matchers';
 import * as urlUtils from './url-utils';
 import {Bundle, BundleStrategy, AssignedBundle, generateBundles, BundleUrlMapper, BundleManifest, sharedBundleUrlMapper, generateSharedDepsMergeStrategy} from './bundle-manifest';
@@ -113,365 +114,6 @@ export class Bundler {
         String(opts.inputUrl) === opts.inputUrl ? opts.inputUrl : '';
   }
 
-
-  /**
-   * Return the URL this import should point to in the given bundle.
-   *
-   * If the URL is part of the bundle, this method returns `true`.
-   *
-   * If the URL is part of another bundle, this method returns the url of that
-   * bundle.
-   *
-   * If the URL isn't part of a bundle, this method returns `false`
-   */
-  resolveBundleUrl(
-      url: string,
-      bundle: AssignedBundle,
-      manifest: BundleManifest): boolean|string {
-    const targetBundle = manifest.getBundleForFile(url);
-    if (!targetBundle || !targetBundle.url) {
-      return false;
-    }
-    if (targetBundle.url !== bundle.url) {
-      const relative = urlUtils.relativeUrl(bundle.url, targetBundle.url);
-      if (!relative) {
-        throw new Error('Unable to compute relative url to bundle');
-      }
-      return relative;
-    }
-    return true;
-  }
-
-  isBlankTextNode(node: ASTNode): boolean {
-    return node && dom5.isTextNode(node) &&
-        /^\s*$/.test(dom5.getTextContent(node));
-  }
-
-  removeElementAndNewline(node: ASTNode, replacement?: ASTNode) {
-    // when removing nodes, remove the newline after it as well
-    const siblings = Array.from(node.parentNode!.childNodes!);
-    let nextIdx = siblings.indexOf(node) + 1;
-    let next = siblings[nextIdx];
-    // remove next node if it is blank text
-    while (next && this.isBlankTextNode(next)) {
-      dom5.remove(next);
-      next = siblings[++nextIdx];
-    }
-    if (replacement) {
-      dom5.replace(node, replacement);
-    } else {
-      dom5.remove(node);
-    }
-  }
-
-  isLicenseComment(node: ASTNode): boolean {
-    if (dom5.isCommentNode(node)) {
-      return dom5.getTextContent(node).indexOf('@license') > -1;
-    }
-    return false;
-  }
-
-  /**
-   * Creates a hidden container <div> to which inlined content will be
-   * appended.
-   */
-  createHiddenDiv(): ASTNode {
-    const hidden = dom5.constructors.element('div');
-    dom5.setAttribute(hidden, 'hidden', '');
-    dom5.setAttribute(hidden, 'by-polymer-bundler', '');
-    return hidden;
-  }
-
-  findOrCreateHiddenDiv(document: ASTNode): ASTNode {
-    const hiddenDiv =
-        dom5.query(document, matchers.hiddenDiv) || this.createHiddenDiv();
-    if (!hiddenDiv.parentNode) {
-      const firstHtmlImport = dom5.query(document, matchers.htmlImport);
-      const body = dom5.query(document, matchers.body);
-      if (body) {
-        if (firstHtmlImport &&
-            dom5.predicates.parentMatches(matchers.body)(firstHtmlImport)) {
-          astUtils.insertAfter(firstHtmlImport, hiddenDiv);
-        } else {
-          astUtils.prepend(body, hiddenDiv);
-        }
-      } else {
-        dom5.append(document, hiddenDiv);
-      }
-    }
-    return hiddenDiv;
-  }
-
-  /**
-   * Inline external scripts <script src="*">
-   */
-  async inlineScript(docUrl: string, externalScript: ASTNode):
-      Promise<ASTNode|undefined> {
-    const rawUrl: string = dom5.getAttribute(externalScript, 'src')!;
-    const resolvedUrl = urlLib.resolve(docUrl, rawUrl);
-    let script: Document|null = null;
-    try {
-      script = await this.analyzer.analyze(resolvedUrl);
-    } catch (err) {
-      // If a script doesn't load, skip inlining.
-      // TODO(garlicnation): use a "canLoad" api on analyzer.
-    }
-
-    if (!script) {
-      return;
-    }
-
-    // Second argument 'true' tells encodeString to escape <script> tags.
-    const scriptContent = encodeString(script.parsedDocument.contents, true);
-    dom5.removeAttribute(externalScript, 'src');
-    dom5.setTextContent(externalScript, scriptContent);
-
-    return externalScript;
-  }
-
-  /**
-   * Inline a stylesheet (either from deprecated polymer-style css import `<link
-   * rel="import" type="css">` import or regular external stylesheet link
-   * `<link rel="stylesheet">`.
-   */
-  async inlineStylesheet(docUrl: string, cssLink: ASTNode):
-      Promise<ASTNode|undefined> {
-    const stylesheetUrl: string = dom5.getAttribute(cssLink, 'href')!;
-    const resolvedStylesheetUrl = urlLib.resolve(docUrl, stylesheetUrl);
-    let stylesheetImport: Document|null = null;
-    try {
-      stylesheetImport = await this.analyzer.analyze(resolvedStylesheetUrl);
-    } catch (err) {
-      // Pass here since there's no canLoad api from the analyzer.
-    }
-
-    if (!stylesheetImport) {
-      return;
-    }
-
-    const media = dom5.getAttribute(cssLink, 'media');
-    const stylesheetContent = stylesheetImport.parsedDocument.contents;
-    const resolvedStylesheetContent = this.rewriteImportedStyleTextUrls(
-        resolvedStylesheetUrl, docUrl, stylesheetContent);
-    const styleNode = dom5.constructors.element('style');
-
-    if (media) {
-      dom5.setAttribute(styleNode, 'media', media);
-    }
-
-    dom5.replace(cssLink, styleNode);
-    dom5.setTextContent(styleNode, resolvedStylesheetContent);
-    return styleNode;
-  }
-
-  /**
-   * Inline external HTML files <link type="import" href="*">
-   * TODO(usergenic): Refactor method to simplify and encapsulate case handling
-   *     for hidden div adjacency etc.
-   */
-  async inlineHtmlImport(
-      docUrl: string,
-      htmlImport: ASTNode,
-      reachedImports: Set<string>,
-      bundle: AssignedBundle,
-      manifest: BundleManifest) {
-    const rawUrl: string = dom5.getAttribute(htmlImport, 'href')!;
-    const resolvedUrl: string = urlLib.resolve(docUrl, rawUrl);
-    const bundleUrl = manifest.bundleUrlForFile.get(resolvedUrl);
-
-    // Don't reprocess the same file again.
-    if (reachedImports.has(resolvedUrl)) {
-      this.removeElementAndNewline(htmlImport);
-      return;
-    }
-
-    // If we can't find a bundle for the referenced import, record that we've
-    // processed it, but don't remove the import link.  Browser will handle it.
-    if (!bundleUrl) {
-      reachedImports.add(resolvedUrl);
-      return;
-    }
-
-    // Don't inline an import into itself.
-    if (docUrl === resolvedUrl) {
-      reachedImports.add(resolvedUrl);
-      this.removeElementAndNewline(htmlImport);
-      return;
-    }
-
-    // Guard against inlining a import we've already processed.
-    if (reachedImports.has(bundleUrl)) {
-      this.removeElementAndNewline(htmlImport);
-      return;
-    }
-
-    // If the html import refers to a file which is bundled and has a different
-    // url, then lets just rewrite the href to point to the bundle url.
-    if (bundleUrl !== bundle.url) {
-      const relative = urlUtils.relativeUrl(docUrl, bundleUrl) || bundleUrl;
-      dom5.setAttribute(htmlImport, 'href', relative);
-      reachedImports.add(bundleUrl);
-      return;
-    }
-
-    const document =
-        dom5.nodeWalkAncestors(htmlImport, (node) => !node.parentNode)!;
-    const body = dom5.query(document, matchers.body)!;
-    const analyzedImport = await this.analyzer.analyze(resolvedUrl);
-
-    // If the document wasn't loaded for the import during analysis, we can't
-    // inline it.
-    if (!analyzedImport) {
-      // TODO(usergenic): What should the behavior be when we don't have the
-      // document to inline available in the analyzer?
-      throw new Error(`Unable to analyze ${resolvedUrl}`);
-    }
-
-    // Is there a better way to get what we want other than using
-    // parseFragment?
-    const importDoc =
-        parse5.parseFragment(analyzedImport.parsedDocument.contents);
-    this.rewriteImportedUrls(importDoc, resolvedUrl, docUrl);
-    const nestedImports = dom5.queryAll(importDoc, matchers.htmlImport);
-
-    // Move all of the import doc content after the html import.
-    astUtils.insertAllBefore(
-        htmlImport.parentNode!, htmlImport, importDoc.childNodes!);
-    this.removeElementAndNewline(htmlImport);
-
-    // If we've never seen this import before, lets add it to the set so we
-    // will deduplicate if we encounter it again.
-    reachedImports.add(resolvedUrl);
-
-    // Recursively process the nested imports.
-    for (const nestedImport of nestedImports) {
-      await this.inlineHtmlImport(
-          docUrl, nestedImport, reachedImports, bundle, manifest);
-    }
-  }
-
-  // TODO(usergenic): Migrate "Old Polymer" detection to polymer-analyzer with
-  // deprecated feature scanners.
-  oldPolymerCheck(analyzedRoot: Document) {
-    analyzedRoot.getByKind('document').forEach((d) => {
-      if (d.parsedDocument instanceof ParsedHtmlDocument &&
-          dom5.query(d.parsedDocument.ast, matchers.polymerElement)) {
-        throw new Error(
-            constants.OLD_POLYMER + ' File: ' + d.parsedDocument.url);
-      }
-    });
-  }
-
-  /**
-   * Given a string of CSS, return a version where all occurrences of urls,
-   * have been rewritten based on the relationship of the import url to the
-   * main doc url.
-   * TODO(usergenic): This is a static method that should probably be moved to
-   * urlUtils or similar.
-   */
-  rewriteImportedStyleTextUrls(
-      importUrl: string,
-      mainDocUrl: string,
-      cssText: string): string {
-    return cssText.replace(constants.URL, match => {
-      let path = match.replace(/["']/g, '').slice(4, -1);
-      path = urlUtils.rewriteImportedRelPath(
-          this.basePath, importUrl, mainDocUrl, path);
-      return 'url("' + path + '")';
-    });
-  }
-
-  rewriteImportedUrls(
-      importDoc: ASTNode,
-      importUrl: string,
-      mainDocUrl: string) {
-    this._rewriteImportedElementAttrUrls(importDoc, importUrl, mainDocUrl);
-    this._rewriteImportedStyleUrls(importDoc, importUrl, mainDocUrl);
-    this._rewriteImportedDomModuleAssetpaths(importDoc, importUrl, mainDocUrl);
-  }
-
-  private _rewriteImportedDomModuleAssetpaths(
-      importDoc: ASTNode,
-      importUrl: string,
-      mainDocUrl: string) {
-    const domModules = dom5.queryAll(importDoc, matchers.domModule);
-    for (let i = 0, node: ASTNode; i < domModules.length; i++) {
-      node = domModules[i];
-      let assetPathUrl = urlUtils.rewriteImportedRelPath(
-          this.basePath, importUrl, mainDocUrl, '');
-      assetPathUrl = pathPosix.dirname(assetPathUrl) + '/';
-      dom5.setAttribute(node, 'assetpath', assetPathUrl);
-    }
-  }
-
-  private _rewriteImportedElementAttrUrls(
-      importDoc: ASTNode,
-      importUrl: string,
-      mainDocUrl: string) {
-    const nodes = dom5.queryAll(importDoc, matchers.urlAttrs);
-    let attrValue: string|null;
-    for (let i = 0, node: ASTNode; i < nodes.length; i++) {
-      node = nodes[i];
-      for (let j = 0, attr: string; j < constants.URL_ATTR.length; j++) {
-        attr = constants.URL_ATTR[j];
-        attrValue = dom5.getAttribute(node, attr);
-        if (attrValue && !urlUtils.isTemplatedUrl(attrValue)) {
-          let relUrl: string;
-          if (attr === 'style') {
-            relUrl = this.rewriteImportedStyleTextUrls(
-                importUrl, mainDocUrl, attrValue);
-          } else {
-            relUrl = urlUtils.rewriteImportedRelPath(
-                this.basePath, importUrl, mainDocUrl, attrValue);
-            if (attr === 'assetpath' && relUrl.slice(-1) !== '/') {
-              relUrl += '/';
-            }
-          }
-          dom5.setAttribute(node, attr, relUrl);
-        }
-      }
-    }
-  }
-
-  private _rewriteImportedStyleUrls(
-      importDoc: ASTNode,
-      importUrl: string,
-      mainDocUrl: string) {
-    const styleNodes = astUtils.querySelectorAllWithTemplates(
-        importDoc, matchers.styleMatcher);
-    for (let i = 0, node: ASTNode; i < styleNodes.length; i++) {
-      node = styleNodes[i];
-      let styleText = dom5.getTextContent(node);
-      styleText =
-          this.rewriteImportedStyleTextUrls(importUrl, mainDocUrl, styleText);
-      dom5.setTextContent(node, styleText);
-    }
-  }
-
-  /**
-   * Old Polymer supported `<style>` tag in `<dom-module>` but outside of
-   * `<template>`.  This is also where the deprecated Polymer CSS import tag
-   * `<link rel="import" type="css">` would generate inline `<style>`.
-   * Migrates these `<style>` tags into available `<template>` of the
-   * `<dom-module>`.  Will create a `<template>` container if not present.
-   */
-  moveDomModuleStyleIntoTemplate(style: ASTNode) {
-    const domModule =
-        dom5.nodeWalkAncestors(style, dom5.predicates.hasTagName('dom-module'));
-    if (!domModule) {
-      // TODO(usergenic): We *shouldn't* get here, but if we do, it's because
-      // the analyzer messed up.
-      return;
-    }
-    let template = dom5.query(domModule, matchers.template);
-    if (!template) {
-      template = dom5.constructors.element('template')!;
-      dom5.append(domModule, template);
-    }
-    this.removeElementAndNewline(style);
-    astUtils.prepend(template, style);
-  }
-
   /**
    * Given a URL to an entry-point html document, produce a single document
    * with HTML imports, external stylesheets and external scripts inlined,
@@ -538,6 +180,157 @@ export class Bundler {
   }
 
   /**
+   * Inline external script content into their tags, converting
+   * `<script src="..."></script>`  tags to `<script>...</script>` tags.
+   */
+  async inlineScript(docUrl: string, externalScript: ASTNode) {
+    return importUtils.inlineScript(
+        docUrl,
+        externalScript,
+        url => this.analyzer.analyze(url).then(
+            doc => doc.parsedDocument.contents));
+  }
+
+  /**
+   * Inline a stylesheet (either from deprecated polymer-style css import `<link
+   * rel="import" type="css">` import or regular external stylesheet link
+   * `<link rel="stylesheet">`.
+   */
+  async inlineStylesheet(docUrl: string, cssLink: ASTNode) {
+    return await importUtils.inlineStylesheet(
+        this.basePath,
+        docUrl,
+        cssLink,
+        url => this.analyzer.analyze(url).then(
+            doc => doc.parsedDocument.contents));
+  }
+
+  /**
+   * Inline external HTML files `<link type="import" href="...">`
+   * TODO(usergenic): Refactor method to simplify and encapsulate case handling
+   *     for hidden div adjacency etc.
+   */
+  async inlineHtmlImport(
+      docUrl: string,
+      htmlImport: ASTNode,
+      reachedImports: Set<string>,
+      bundle: AssignedBundle,
+      manifest: BundleManifest) {
+    const rawUrl: string = dom5.getAttribute(htmlImport, 'href')!;
+    const resolvedUrl: string = urlLib.resolve(docUrl, rawUrl);
+    const bundleUrl = manifest.bundleUrlForFile.get(resolvedUrl);
+
+    // Don't reprocess the same file again.
+    if (reachedImports.has(resolvedUrl)) {
+      astUtils.removeElementAndNewline(htmlImport);
+      return;
+    }
+
+    // If we can't find a bundle for the referenced import, record that we've
+    // processed it, but don't remove the import link.  Browser will handle it.
+    if (!bundleUrl) {
+      reachedImports.add(resolvedUrl);
+      return;
+    }
+
+    // Don't inline an import into itself.
+    if (docUrl === resolvedUrl) {
+      reachedImports.add(resolvedUrl);
+      astUtils.removeElementAndNewline(htmlImport);
+      return;
+    }
+
+    // Guard against inlining a import we've already processed.
+    if (reachedImports.has(bundleUrl)) {
+      astUtils.removeElementAndNewline(htmlImport);
+      return;
+    }
+
+    // If the html import refers to a file which is bundled and has a different
+    // url, then lets just rewrite the href to point to the bundle url.
+    if (bundleUrl !== bundle.url) {
+      const relative = urlUtils.relativeUrl(docUrl, bundleUrl) || bundleUrl;
+      dom5.setAttribute(htmlImport, 'href', relative);
+      reachedImports.add(bundleUrl);
+      return;
+    }
+
+    const document =
+        dom5.nodeWalkAncestors(htmlImport, (node) => !node.parentNode)!;
+    const body = dom5.query(document, matchers.body)!;
+    const analyzedImport = await this.analyzer.analyze(resolvedUrl);
+
+    // If the document wasn't loaded for the import during analysis, we can't
+    // inline it.
+    if (!analyzedImport) {
+      // TODO(usergenic): What should the behavior be when we don't have the
+      // document to inline available in the analyzer?
+      throw new Error(`Unable to analyze ${resolvedUrl}`);
+    }
+
+    // Is there a better way to get what we want other than using
+    // parseFragment?
+    const importDoc =
+        parse5.parseFragment(analyzedImport.parsedDocument.contents);
+    importUtils.rewriteImportedUrls(
+        this.basePath, importDoc, resolvedUrl, docUrl);
+    const nestedImports = dom5.queryAll(importDoc, matchers.htmlImport);
+
+    // Move all of the import doc content after the html import.
+    astUtils.insertAllBefore(
+        htmlImport.parentNode!, htmlImport, importDoc.childNodes!);
+    astUtils.removeElementAndNewline(htmlImport);
+
+    // If we've never seen this import before, lets add it to the set so we
+    // will deduplicate if we encounter it again.
+    reachedImports.add(resolvedUrl);
+
+    // Recursively process the nested imports.
+    for (const nestedImport of nestedImports) {
+      await this.inlineHtmlImport(
+          docUrl, nestedImport, reachedImports, bundle, manifest);
+    }
+  }
+
+  /**
+   * Check the document for references to the old `<polymer-element>` custom
+   * element.  If one is found, throw an error calling it out.
+   * TODO(usergenic): Migrate "Old Polymer" detection to polymer-analyzer with
+   * deprecated feature scanners.
+   * TODO(usergenic): This behavior is a bit severe and would prevent developers
+   * from using their own custom element named polymer-element, which they may
+   * choose to do for unrelated reasons.
+   */
+  oldPolymerCheck(analyzedRoot: Document) {
+    analyzedRoot.getByKind('document').forEach((d) => {
+      if (d.parsedDocument instanceof ParsedHtmlDocument &&
+          dom5.query(d.parsedDocument.ast, matchers.polymerElement)) {
+        throw new Error(
+            constants.OLD_POLYMER + ' File: ' + d.parsedDocument.url);
+      }
+    });
+  }
+
+  /**
+   * Add HTML Import elements for each file in the bundle.  We append all the
+   * imports in the case any were moved into the bundle by the strategy.
+   * While this will almost always yield duplicate imports, they will be
+   * cleaned up through deduplication during the import phase.
+   */
+  private _appendHtmlImportsForBundle(
+      document: ASTNode,
+      bundle: AssignedBundle) {
+    for (const importUrl of bundle.bundle.files) {
+      const newUrl = urlUtils.relativeUrl(bundle.url, importUrl);
+      if (!newUrl) {
+        continue;
+      }
+      this._appendImport(this._findOrCreateHiddenDiv(document), newUrl);
+    }
+    return document;
+  }
+
+  /**
    * Append a <link rel="import" node to `node` with a value of `url` for
    * the "href" attribute.
    */
@@ -547,6 +340,111 @@ export class Bundler {
     dom5.setAttribute(newNode, 'href', url);
     dom5.append(node, newNode);
     return newNode;
+  }
+
+  /**
+   * Set the hidden div at the appropriate location within the document.  The
+   * goal is to place the hidden div at the same place as the first html
+   * import.  However, the div can't be placed in the `<head>` of the document
+   * so if first import is found in the head, we prepend the div to the body.
+   * If there is no body, we'll just attach the hidden div to the document at
+   * the end.
+   */
+  private _attachHiddenDiv(document: ASTNode, hiddenDiv: ASTNode) {
+    const firstHtmlImport = dom5.query(document, matchers.htmlImport);
+    const body = dom5.query(document, matchers.body);
+    if (body) {
+      if (firstHtmlImport &&
+          dom5.predicates.parentMatches(matchers.body)(firstHtmlImport)) {
+        astUtils.insertAfter(firstHtmlImport, hiddenDiv);
+      } else {
+        astUtils.prepend(body, hiddenDiv);
+      }
+    } else {
+      dom5.append(document, hiddenDiv);
+    }
+  }
+
+  /**
+   * TODO(garlicnation): resolve <base> tags.
+   * TODO(garlicnation): find transitive dependencies of specified excluded
+   * files.
+   * TODO(garlicnation): ignore <link> in <template>
+   * TODO(garlicnation): Support addedImports
+   *
+   * SAVED FROM buildLoader COMMENTS
+   * TODO(garlicnation): Add noopResolver for external urls.
+   * TODO(garlicnation): Add redirectResolver for fakeprotocol:// urls
+   * TODO(garlicnation): Add noopResolver for excluded urls.
+   */
+  private async _bundleDocument(
+      bundle: AssignedBundle,
+      bundleManifest: BundleManifest,
+      bundleImports?: Set<string>): Promise<ASTNode> {
+    const url = bundle.url;
+    const document = await this._prepareBundleDocument(bundle);
+    this._appendHtmlImportsForBundle(document, bundle);
+    let analyzedRoot: any;
+    try {
+      analyzedRoot =
+          await this.analyzer.analyze(url, parse5.serialize(document));
+    } catch (err) {
+      throw new Error('Unable to analyze document!');
+    }
+
+    const head: ASTNode = dom5.query(document, matchers.head)!;
+    const body: ASTNode = dom5.query(document, matchers.body)!;
+
+    const elementInHead = dom5.predicates.parentMatches(matchers.head);
+
+    importUtils.rewriteImportedUrls(this.basePath, document, url, url);
+
+    // Old Polymer versions are not supported, so warn user.
+    this.oldPolymerCheck(analyzedRoot);
+
+    const reachedImports = new Set<UrlString>();
+
+    // Inline all HTML Imports, using "reachedImports" for deduplication.
+    await this._inlineHtmlImports(url, document, bundle, bundleManifest);
+
+    if (this.enableScriptInlining) {
+      await this._inlineScripts(url, document);
+    }
+
+    if (this.enableCssInlining) {
+      await this._inlineStylesheetLinks(url, document);
+      await this._inlineStylesheetImports(url, document);
+    }
+
+    if (this.stripComments) {
+      this._stripComments(document);
+    }
+    return document;
+  }
+
+  /**
+   * Creates a hidden container <div> to which inlined content will be
+   * appended.
+   */
+  private _createHiddenDiv(): ASTNode {
+    const hidden = dom5.constructors.element('div');
+    dom5.setAttribute(hidden, 'hidden', '');
+    dom5.setAttribute(hidden, 'by-polymer-bundler', '');
+    return hidden;
+  }
+
+  /**
+   * Given a document, search for the hidden div, if it isn't found, then
+   * create it.  After creating it, attach it to the desired location.  Then
+   * return it.
+   */
+  private _findOrCreateHiddenDiv(document: ASTNode): ASTNode {
+    const hiddenDiv =
+        dom5.query(document, matchers.hiddenDiv) || this._createHiddenDiv();
+    if (!hiddenDiv.parentNode) {
+      this._attachHiddenDiv(document, hiddenDiv);
+    }
+    return hiddenDiv;
   }
 
   private async _inlineHtmlImports(
@@ -583,7 +481,7 @@ export class Bundler {
     for (const cssLink of cssImports) {
       const style = await this.inlineStylesheet(url, cssLink);
       if (style) {
-        this.moveDomModuleStyleIntoTemplate(style);
+        this._moveDomModuleStyleIntoTemplate(style);
       }
     }
   }
@@ -600,6 +498,29 @@ export class Bundler {
     }
   }
 
+  /**
+   * Old Polymer supported `<style>` tag in `<dom-module>` but outside of
+   * `<template>`.  This is also where the deprecated Polymer CSS import tag
+   * `<link rel="import" type="css">` would generate inline `<style>`.
+   * Migrates these `<style>` tags into available `<template>` of the
+   * `<dom-module>`.  Will create a `<template>` container if not present.
+   */
+  private _moveDomModuleStyleIntoTemplate(style: ASTNode) {
+    const domModule =
+        dom5.nodeWalkAncestors(style, dom5.predicates.hasTagName('dom-module'));
+    if (!domModule) {
+      // TODO(usergenic): We *shouldn't* get here, but if we do, it's because
+      // the analyzer messed up.
+      return;
+    }
+    let template = dom5.query(domModule, matchers.template);
+    if (!template) {
+      template = dom5.constructors.element('template')!;
+      dom5.append(domModule, template);
+    }
+    astUtils.removeElementAndNewline(style);
+    astUtils.prepend(template, style);
+  }
   /**
    * When an HTML Import is encountered in the head of the document, it needs
    * to be moved into the hidden div and any subsequent order-dependent
@@ -618,8 +539,8 @@ export class Bundler {
     for (const node of [firstHtmlImport].concat(
              astUtils.siblingsAfter(firstHtmlImport))) {
       if (matchers.orderedImperative(node)) {
-        this.removeElementAndNewline(node);
-        dom5.append(this.findOrCreateHiddenDiv(document), node);
+        astUtils.removeElementAndNewline(node);
+        dom5.append(this._findOrCreateHiddenDiv(document), node);
       }
     }
   }
@@ -634,8 +555,8 @@ export class Bundler {
         dom5.predicates.AND(
             matchers.htmlImport, dom5.predicates.NOT(matchers.inHiddenDiv)));
     for (const htmlImport of unhiddenHtmlImports) {
-      this.removeElementAndNewline(htmlImport);
-      dom5.append(this.findOrCreateHiddenDiv(document), htmlImport);
+      astUtils.removeElementAndNewline(htmlImport);
+      dom5.append(this._findOrCreateHiddenDiv(document), htmlImport);
     }
   }
 
@@ -667,95 +588,13 @@ export class Bundler {
     dom5.nodeWalkAll(document, dom5.isCommentNode)
         .forEach((comment: ASTNode) => {
           comments.set(comment.data || '', comment);
-          this.removeElementAndNewline(comment);
+          astUtils.removeElementAndNewline(comment);
         });
     const head = dom5.query(document, matchers.head);
     for (const comment of comments.values()) {
-      if (this.isLicenseComment(comment)) {
+      if (astUtils.isLicenseComment(comment)) {
         astUtils.prepend(head || document, comment);
       }
     }
-  }
-
-  private async _synthesizeBundleContents(
-      bundle: AssignedBundle,
-      reachedImports: Set<UrlString>) {
-    const document = await this._prepareBundleDocument(bundle);
-    const body = dom5.query(document, matchers.body);
-    if (!body) {
-      throw new Error('Unexpected return from parse5.parse');
-    }
-
-    // Add HTML Import elements for each file in the bundle.  We append all the
-    // imports in the case any were moved into the bundle by the strategy.
-    // While this will almost always yield duplicate imports, they will be
-    // cleaned up through deduplication during the import phase.
-    for (const importUrl of bundle.bundle.files) {
-      const newUrl = urlUtils.relativeUrl(bundle.url, importUrl);
-      if (!newUrl) {
-        continue;
-      }
-      this._appendImport(this.findOrCreateHiddenDiv(document), newUrl);
-    }
-    return document;
-  }
-
-  private async _bundleDocument(
-      bundle: AssignedBundle,
-      bundleManifest: BundleManifest,
-      bundleImports?: Set<string>): Promise<ASTNode> {
-    const url = bundle.url;
-    // Set tracking imports that have been reached.
-    const inlinedImports: Set<UrlString> = new Set();
-    const document =
-        await this._synthesizeBundleContents(bundle, inlinedImports);
-    let analyzedRoot: any;
-    try {
-      analyzedRoot =
-          await this.analyzer.analyze(url, parse5.serialize(document));
-    } catch (err) {
-      throw new Error('Unable to analyze document!');
-    }
-
-    const head: ASTNode = dom5.query(document, matchers.head)!;
-    const body: ASTNode = dom5.query(document, matchers.body)!;
-
-    const elementInHead = dom5.predicates.parentMatches(matchers.head);
-
-    this.rewriteImportedUrls(document, url, url);
-
-    // Old Polymer versions are not supported, so warn user.
-    this.oldPolymerCheck(analyzedRoot);
-
-    const reachedImports = new Set<UrlString>();
-
-    // Inline all HTML Imports, using "reachedImports" for deduplication.
-    await this._inlineHtmlImports(url, document, bundle, bundleManifest);
-
-    if (this.enableScriptInlining) {
-      await this._inlineScripts(url, document);
-    }
-
-    if (this.enableCssInlining) {
-      await this._inlineStylesheetLinks(url, document);
-      await this._inlineStylesheetImports(url, document);
-    }
-
-    if (this.stripComments) {
-      this._stripComments(document);
-    }
-    return document;
-
-    // LATER
-    // TODO(garlicnation): resolve <base> tags.
-    // TODO(garlicnation): find transitive dependencies of specified excluded
-    // files.
-    // TODO(garlicnation): ignore <link> in <template>
-    // TODO(garlicnation): Support addedImports
-
-    // SAVED FROM buildLoader COMMENTS
-    // TODO(garlicnation): Add noopResolver for external urls.
-    // TODO(garlicnation): Add redirectResolver for fakeprotocol:// urls
-    // TODO(garlicnation): Add noopResolver for excluded urls.
   }
 }
